+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -92,7 +92,7 @@
   - title: State
     url: "/components/beta/state"
   - title: Subhead
-    url: "/components/subhead"
+    url: "/components/beta/subhead"
   - title: TabContainer
     url: "/components/tabcontainer"
   - title: TabNav
