+++ lib/primer/view_components/linters/subhead_component_migration_counter.rb
@@ -6,7 +6,7 @@ module ERBLint
   module Linters
     # Counts the number of times a HTML Subhead is used instead of the component.
     class SubheadComponentMigrationCounter < BaseLinter
-      MESSAGE = "We are migrating Subhead to use [Primer::SubheadComponent](https://primer.style/view-components/components/subhead), please try to use that instead of raw HTML."
+      MESSAGE = "We are migrating Subhead to use [Primer::Beta::Subhead](https://primer.style/view-components/components/subhead), please try to use that instead of raw HTML."
       CLASSES = %w[Subhead].freeze
       TAGS = %w[div].freeze
     end
