+++ static/audited_at.json
@@ -63,6 +63,7 @@
   "Primer::Beta::RelativeTime": "",
   "Primer::Beta::Spinner": "",
   "Primer::Beta::State": "",
+  "Primer::Beta::Subhead": "",
   "Primer::Beta::Text": "",
   "Primer::Beta::Truncate": "",
   "Primer::Beta::Truncate::TruncateText": "",
