+++ static/constants.json
@@ -958,6 +958,18 @@
       "div"
     ]
   },
+  "Primer::Beta::Subhead": {
+    "DEFAULT_HEADING_TAG": "div",
+    "HEADING_TAG_OPTIONS": [
+      "div",
+      "h1",
+      "h2",
+      "h3",
+      "h4",
+      "h5",
+      "h6"
+    ]
+  },
   "Primer::Beta::Text": {
     "DEFAULT_TAG": "span"
   },
@@ -1104,16 +1116,6 @@
   "Primer::StateComponent": {
   },
   "Primer::SubheadComponent": {
-    "DEFAULT_HEADING_TAG": "div",
-    "HEADING_TAG_OPTIONS": [
-      "div",
-      "h1",
-      "h2",
-      "h3",
-      "h4",
-      "h5",
-      "h6"
-    ]
   },
   "Primer::TabContainerComponent": {
   },
