+++ static/statuses.json
@@ -63,6 +63,7 @@
   "Primer::Beta::RelativeTime": "beta",
   "Primer::Beta::Spinner": "beta",
   "Primer::Beta::State": "beta",
+  "Primer::Beta::Subhead": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
   "Primer::Beta::Truncate::TruncateText": "alpha",
@@ -91,7 +92,7 @@
   "Primer::PopoverComponent": "deprecated",
   "Primer::SpinnerComponent": "deprecated",
   "Primer::StateComponent": "deprecated",
-  "Primer::SubheadComponent": "beta",
+  "Primer::SubheadComponent": "deprecated",
   "Primer::TabContainerComponent": "alpha",
   "Primer::TimeAgoComponent": "deprecated",
   "Primer::TimelineItemComponent": "beta",
