+++ test/components/beta/subhead_test.rb
@@ -2,17 +2,17 @@
 
 require "components/test_helper"
 
-class SubheadComponentTest < Minitest::Test
+class PrimerBetaSubheadTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_does_not_render_without_heading
-    render_inline(Primer::SubheadComponent.new)
+    render_inline(Primer::Beta::Subhead.new)
 
     refute_component_rendered
   end
 
   def test_renders_heading
-    render_inline(Primer::SubheadComponent.new) do |component|
+    render_inline(Primer::Beta::Subhead.new) do |component|
       component.with_heading(tag: :h2) { "Hello World" }
     end
 
@@ -21,7 +21,7 @@ class SubheadComponentTest < Minitest::Test
 
   def test_heading_tag_falls_back_to_default
     without_fetch_or_fallback_raises do
-      render_inline(Primer::SubheadComponent.new) do |component|
+      render_inline(Primer::Beta::Subhead.new) do |component|
         component.with_heading(tag: :span) { "Hello World" }
       end
     end
@@ -30,7 +30,7 @@ class SubheadComponentTest < Minitest::Test
   end
 
   def test_render_dangerous_heading
-    render_inline(Primer::SubheadComponent.new) do |component|
+    render_inline(Primer::Beta::Subhead.new) do |component|
       component.with_heading(danger: true) { "Hello World" }
     end
 
@@ -38,7 +38,7 @@ class SubheadComponentTest < Minitest::Test
   end
 
   def test_render_without_border
-    render_inline(Primer::SubheadComponent.new(hide_border: true)) do |component|
+    render_inline(Primer::Beta::Subhead.new(hide_border: true)) do |component|
       component.with_heading { "Hello World" }
     end
 
@@ -46,7 +46,7 @@ class SubheadComponentTest < Minitest::Test
   end
 
   def test_bottom_margin_can_be_overridden_when_border_is_hidden
-    render_inline(Primer::SubheadComponent.new(hide_border: true, mb: 1)) do |component|
+    render_inline(Primer::Beta::Subhead.new(hide_border: true, mb: 1)) do |component|
       component.with_heading { "Hello World" }
     end
 
@@ -54,7 +54,7 @@ class SubheadComponentTest < Minitest::Test
   end
 
   def test_renders_actions
-    render_inline(Primer::SubheadComponent.new(heading: "Hello world")) do |component|
+    render_inline(Primer::Beta::Subhead.new(heading: "Hello world")) do |component|
       component.with_heading { "Hello World" }
       component.with_actions { "My Actions" }
     end
@@ -63,7 +63,7 @@ class SubheadComponentTest < Minitest::Test
   end
 
   def test_handles_spacious
-    render_inline(Primer::SubheadComponent.new(spacious: true)) do |component|
+    render_inline(Primer::Beta::Subhead.new(spacious: true)) do |component|
       component.with_heading { "Hello World" }
     end
 
@@ -71,7 +71,7 @@ class SubheadComponentTest < Minitest::Test
   end
 
   def test_renders_a_description
-    render_inline(Primer::SubheadComponent.new(heading: "Hello world")) do |component|
+    render_inline(Primer::Beta::Subhead.new(heading: "Hello world")) do |component|
       component.with_heading { "Hello World" }
       component.with_description { "My Description" }
     end
@@ -80,6 +80,6 @@ class SubheadComponentTest < Minitest::Test
   end
 
   def test_status
-    assert_component_state(Primer::SubheadComponent, :beta)
+    assert_component_state(Primer::Beta::Subhead, :beta)
   end
 end
